move weather logic to location
