+++ src/pages/Location/Location.js
@@ -18,15 +18,19 @@ import React, {
 } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { Search } from '@mui/icons-material';
+import Axios from 'axios';
+import moment from 'moment';
 import { Map } from '@psa/dst.ui.map';
 // import centroid from '@turf/centroid';
 import mapboxgl from 'mapbox-gl';
 import statesLatLongDict from '../../shared/stateslatlongdict';
-import { BinaryButton } from '../../shared/constants';
+import { abbrRegion, reverseGEO, BinaryButton } from '../../shared/constants';
 import { Context } from '../../store/Store';
 import MyCoverCropReset from '../../components/MyCoverCropReset/MyCoverCropReset';
 import PlantHardinessZone from '../CropSidebar/PlantHardinessZone/PlantHardinessZone';
-import { changeAddressViaMap, updateLocation, updateZone as updateZoneRedux } from '../../reduxStore/addressSlice';
+import {
+  changeAddressViaMap, updateLocation, updateZone as updateZoneRedux, updateZipCode,
+} from '../../reduxStore/addressSlice';
 
 // eslint-disable-next-line import/no-webpack-loader-syntax, import/no-unresolved
 mapboxgl.workerClass = require('worker-loader!mapbox-gl/dist/mapbox-gl-csp-worker').default;
@@ -44,6 +48,7 @@ const LocationComponent = ({
   const defaultMarkers = [[40.78489145, -74.80733626930342]];
   const countyRedux = useSelector((stateRedux) => stateRedux.addressData.county);
   const zoneRedux = useSelector((stateRedux) => stateRedux.addressData.zone);
+  const markersRedux = useSelector((stateRedux) => stateRedux.addressData.markers);
 
   const getLatLng = useCallback(() => {
     if (state.state) {
@@ -191,6 +196,162 @@ const LocationComponent = ({
       }
     }
   }, [selectedToEditSite]);
+
+  useEffect(() => {
+    // const { markers } = state;
+    const weatherApiURL = 'https://weather.covercrop-data.org';
+
+    // update address on marker change
+    // ref forecastComponent
+    const lat = markersRedux[0][0];
+    const lon = markersRedux[0][1];
+
+    // since this updates with state; ideally, weather and soil info should be updated here
+    // get current lat long and get county, state and city
+    if (state.progress >= 1 && markersRedux.length > 0) {
+      reverseGEO(lat, lon)
+        .then(async (resp) => {
+          const abbrState = abbrRegion(
+            resp?.features?.filter((feature) => feature?.place_type?.includes('region'))[0]?.text,
+            'abbr',
+          ).toLowerCase();
+
+          const city = resp?.features?.filter((feature) => feature?.place_type?.includes('place'))[0]?.text?.toLowerCase();
+          const zip = resp?.features?.filter((feature) => feature?.place_type?.includes('postcode'))[0]?.text;
+
+          if (zip) {
+            dispatchRedux(updateZipCode(zip));
+          }
+
+          const frostUrl = `${weatherApiURL}/frost?lat=${lat}&lon=${lon}`;
+          await Axios.get(frostUrl)
+            .then(((frostResp) => {
+              // added "/" and do %100 to get them into correct format (want frost dates to look like 01/01/23)
+              const currYear = `/${(new Date().getFullYear() % 100).toString()}`;
+              const prevYear = `/${((new Date().getFullYear() % 100) - 1).toString()}`;
+              const oneDay = 24 * 60 * 60 * 1000; // milliseconds in a day
+              const firstFrost = new Date(frostResp.data.firstfrost + prevYear);
+              const lastFrost = new Date(frostResp.data.lastfrost + currYear);
+
+              const frostFreeDaysObj = Math.round(Math.abs((firstFrost.valueOf() - lastFrost.valueOf()) / oneDay));
+              const averageFrostObject = {
+                firstFrostDate: {
+                  month: firstFrost.toLocaleString('en-US', { month: 'long' }),
+                  day: firstFrost.getDate().toString(),
+                },
+                lastFrostDate: {
+                  month: lastFrost.toLocaleString('en-US', { month: 'long' }),
+                  day: lastFrost.getDate().toString(),
+                },
+              };
+
+              return {
+                frostFreeDaysObj, city, abbrState, averageFrostObject,
+              };
+            }))
+            .then((obj) => {
+              dispatch({
+                type: 'UPDATE_FROST_FREE_DAYS',
+                data: { frostFreeDays: obj.frostFreeDaysObj },
+              });
+              dispatch({
+                type: 'UPDATE_AVERAGE_FROST_DATES',
+                data: {
+                  averageFrost: obj.averageFrostObject,
+                },
+              });
+              return obj;
+            })
+            .then(async (obj) => {
+              const currentMonthInt = moment().month() + 1;
+
+              // What was the 5-year average rainfall for city st during the month of currentMonthInt?
+              //  Dynamic dates ?
+              const averageRainUrl = `${weatherApiURL}/hourly?location=${obj.city}%20${obj.abbrState}&start=2015-01-01&end=2019-12-31`;
+              const averageRainForAMonthURL = `${averageRainUrl}&stats=sum(precipitation)/5&where=month=${currentMonthInt}&output=json`;
+              // What was the 5-year average annual rainfall for city st?
+              const fiveYearAvgRainURL = `${averageRainUrl}&stats=sum(precipitation)/5&output=json`;
+              if (!abbrState.ajaxInProgress) {
+                dispatch({
+                  type: 'SET_AJAX_IN_PROGRESS',
+                  data: true,
+                });
+                await Axios.get(averageRainForAMonthURL)
+                  .then((rainResp) => {
+                    let averagePrecipitationForCurrentMonth = rainResp.data[0]['sum(precipitation)/5'];
+                    averagePrecipitationForCurrentMonth = parseFloat(
+                      averagePrecipitationForCurrentMonth,
+                    ).toFixed(2);
+                    averagePrecipitationForCurrentMonth = parseFloat(
+                      averagePrecipitationForCurrentMonth * 0.03937,
+                    ).toFixed(2);
+                    dispatch({
+                      type: 'UPDATE_AVERAGE_PRECIP_CURRENT_MONTH',
+                      data: { thisMonth: averagePrecipitationForCurrentMonth },
+                    });
+                  })
+                  .catch((error) => {
+                    dispatch({
+                      type: 'SNACK',
+                      data: {
+                        snackOpen: true,
+                        snackMessage: `Weather API error code: ${error.response.status} for getting 5 year average rainfall for this month`,
+                      },
+                    });
+                  });
+
+                if (!abbrState.ajaxInProgress) {
+                  dispatch({
+                    type: 'SET_AJAX_IN_PROGRESS',
+                    data: true,
+                  });
+                  await Axios.get(fiveYearAvgRainURL)
+                    .then((rainResp) => {
+                      let fiveYearAvgRainAnnual = rainResp.data[0]['sum(precipitation)/5'];
+                      fiveYearAvgRainAnnual = parseFloat(fiveYearAvgRainAnnual).toFixed(2);
+                      fiveYearAvgRainAnnual = parseFloat(fiveYearAvgRainAnnual * 0.03937).toFixed(
+                        2,
+                      );
+                      dispatch({
+                        type: 'UPDATE_AVERAGE_PRECIP_ANNUAL',
+                        data: { annual: fiveYearAvgRainAnnual },
+                      });
+                      dispatch({
+                        type: 'SET_AJAX_IN_PROGRESS',
+                        data: false,
+                      });
+                    })
+                    .then(() => { })
+                    .catch((error) => {
+                      dispatch({
+                        type: 'SNACK',
+                        data: {
+                          snackOpen: true,
+                          snackMessage: `Weather API error code: ${error.response.status
+                          } for getting 5 year average rainfall for ${obj.city.toUpperCase()}, ${obj.state.toUpperCase()}`,
+                        },
+                      });
+                      dispatch({
+                        type: 'SET_AJAX_IN_PROGRESS',
+                        data: false,
+                      });
+                    });
+                }
+              }
+            })
+            .catch((err) => {
+              // eslint-disable-next-line no-console
+              console.error(`Failed to fetch frost data: ${err}`);
+            });
+        })
+        .then(() => {
+          dispatch({
+            type: 'SET_AJAX_IN_PROGRESS',
+            data: false,
+          });
+        });
+    }
+  }, [markersRedux]);
 
   return (
     <div className="container-fluid mt-5">
